requirements.txt
aiofiles==0.6.0
aiohttp==3.6.3
aiohttp-swagger==1.0.15
ansible==3.0.0
ansible-base==2.10.6
arrow==1.0.2
async-lru==1.0.2
async-timeout==3.0.1
attrs==20.3.0
backports.ssl==0.0.9
binaryornot==0.4.4
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorful==0.5.4
cookiecutter==1.7.2
cryptography==3.3.1
dill==0.3.3
distro==1.5.0
dnspython==1.16.0
genie==21.2
genie.libs.clean==21.2
genie.libs.conf==21.2
genie.libs.filetransferutils==21.2
genie.libs.health==21.2
genie.libs.ops==21.2
genie.libs.parser==21.2
genie.libs.robot==21.2
genie.libs.sdk==21.2
genie.telemetry==21.2
genie.trafficgen==21.2
idna==2.10
IxNetwork==9.0.1915.16
Jinja2==2.11.2
jinja2-time==0.2.0
jsonpickle==2.0.0
junit-xml==1.9
MarkupSafe==1.1.1
msgpack==1.0.2
multidict==4.7.6
netaddr==0.8.0
packaging==20.9
pathspec==0.8.1
ply==3.11
poyo==0.5.0
prettyprinter==0.18.0
prettytable==2.0.0
psutil==5.8.0
pyasn1==0.4.8
pyats==21.2
pyats.aereport==21.2
pyats.aetest==21.2
pyats.async==21.2
pyats.connections==21.2.1
pyats.contrib==21.2
pyats.datastructures==21.2
pyats.easypy==21.2
pyats.kleenex==21.2
pyats.log==21.2
pyats.reporter==21.2
pyats.results==21.2
pyats.robot==21.2
pyats.tcl==21.2
pyats.topology==21.2
pyats.utils==21.2
pycparser==2.20
pycryptodomex==3.10.1
pyftpdlib==1.5.6
Pygments==2.8.0
pymongo==3.11.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==3.13.2
python-slugify==4.0.1
python-socketio==4.6.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
robotframework==3.2.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Rx==3.1.1
schedule==1.0.0
six==1.15.0
text-unidecode==1.3
tftpy==0.8.0
tqdm==4.58.0
unicon==21.2
unicon.plugins==21.2
urllib3==1.26.3
watchdog==2.0.1
wcwidth==0.2.5
websocket-client==0.57.0
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
xmltodict==0.12.0
yamllint==1.26.0
yarl==1.5.1
